Move the LoadFromTemplate helper to a separate package
This will allow both vtsql and vtctldclient to access it
